se corrigio la variable h

+++ Modelo_EOQ_Clasico.java
@@ -7,14 +7,14 @@ public class Modelo_EOQ_Clasico{
         System.out.println("Ingrese el costo de la demanda: ");
         int k = lec.nextInt();
         System.out.println("Ingrese el costo de almacen del producto: ");
-        int h = lec.nextInt();
+        double h = lec.nextInt();
         int y = 0;
         double resultado = Operaciones(d,k,h, y);
         double entrega = Tiempo(y, d);
         System.out.println("El tamaño optimo del pedido es: " + resultado);
         System.out.println("El tiempo entre pedidos es: " + entrega);
     }
-    public static double Operaciones(int d, int k, int h, int y){
+    public static double Operaciones(int d, int k, double h, int y){
         return y = (int) Math.sqrt(2*k*d/h);
     }
     public static double Tiempo(int y, int d){
